pyproject.toml
[tool.poetry]
name = "vndb"
version = "0.1.0"
description = ""
authors = ["andrey <kuznetsow42@gmail.com>"]
readme = "README.md"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "vnd.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
django-filter = "^23.3"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.0"
djangorestframework-serializer-extensions = "^2.0.1"
nh3 = "^0.2.14"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
